initial

+++ spp-portfolio-constituents-manager/src/main/java/spp/portfolio/constituents/rules/inmemory/PortfolioConfiguration.java
@@ -1,12 +1,13 @@
 package spp.portfolio.constituents.rules.inmemory;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
-import spp.portfolio.model.definition.configuration.constituents.PortfolioConfigurationConstituents;
 
 @Data
 public class PortfolioConfiguration
 {
     private final Currency currency;
-    private final Long portfolioAmountLimit;
+    private final BigDecimal portfolioAmountLimit;
     private PortfolioConfigurationConstituents portfolioConfigurationConstituents;
 }
