+++ spp-portfolio-model/src/main/java/spp/portfolio/model/definition/configuration/PortfolioConfiguration.java
@@ -1,5 +1,7 @@
 package spp.portfolio.model.definition.configuration;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 import spp.portfolio.model.definition.configuration.constituents.PortfolioConfigurationConstituents;
 
@@ -7,6 +9,6 @@ import spp.portfolio.model.definition.configuration.constituents.PortfolioConfig
 public class PortfolioConfiguration
 {
     private final Currency currency;
-    private final Long portfolioAmountLimit;
+    private final BigDecimal portfolioAmountLimit;
     private PortfolioConfigurationConstituents portfolioConfigurationConstituents;
 }
